Revert "se corrige funcion delete en centro de costos y se corrigen querys"
This reverts commit 8e410e23bf09ebf0dcc44b7dcdb8ca64d8fa8cb2.

+++ controllers/CostCenterController.js
@@ -356,10 +356,10 @@ export default class CostCenterController {
     }
 
     async delete(req,res){
-        // const hasAccess = await this.validatePermission(req.body);
-        // if(!hasAccess.ok){
-        //     return res.json(hasAccess)
-        // }
+        const hasAccess = await this.validatePermission(req.body);
+        if(!hasAccess.ok){
+            return res.json(hasAccess)
+        }
 
         try {
             await this.costCenterService.delete(req.params['id']);
